+++ WebMarket/Views/Employees/Details.cshtml
@@ -23,7 +23,7 @@
 
     <p>
 
-        <a asp-controller="Home" asp-action="Employees">Назад к списку</a>
+        <a asp-controller="Employees" asp-action="Employees">Назад к списку</a>
 
     </p>
